pyproject.toml
[project]
name = "turtle-canslim"
version = "0.1.0"
description = "CANSLIM + Turtle Trading automated stock trading system"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Private"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = ["trading", "stocks", "canslim", "turtle-trading", "automation"]

dependencies = [
    # 한투 API
    "python-kis>=2.1.0",
    
    # 데이터 처리
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    
    # 데이터베이스
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "alembic>=1.13.0",
    
    # 캐시
    "redis>=5.0.0",
    
    # HTTP
    "httpx>=0.25.0",
    
    # 스케줄링
    "apscheduler>=3.10.0",
    
    # 알림
    "python-telegram-bot>=20.0",
    
    # 설정 & 검증
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    
    # 유틸리티
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    
    # TUI
    "textual>=0.47.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    # 테스트
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    
    # 타입 체크
    "mypy>=1.7.0",
    "pandas-stubs>=2.0.0",
    "types-PyYAML>=6.0.0",
    "types-redis>=4.6.0",
    
    # 린팅
    "ruff>=0.1.0",
    
    # 백테스트
    "backtesting>=0.3.3",
    "vectorbt>=0.25.0",
]

[project.scripts]
turtle-tui = "scripts.run_tui:main"
turtle-screen = "scripts.run_screener:main"
turtle-trade = "scripts.run_trading:main"
turtle-backtest = "scripts.run_backtest:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG"]  # Unused arguments OK in tests

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "python_kis.*",
    "backtesting.*",
    "vectorbt.*",
    "apscheduler.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "backtest: Backtest tests",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
